- Addressed more code analysis warnings related to int->size_t, this time focused on converting the CGrayMulti and CGrayStaticsBlock classes to use the 'correct' data types. - Added several static_cast<>s in header files where we're assigning enum values to smaller data types. This is primarily to avoid the mass of compiler warnings that get in the way of the important ones, and have been marked with a 'future:' comment so these casts can be revisited once compilers eventually support specifying the correct underlying data types. - Added a CResource::GetMultiItemDefs overload which takes a CItem* and forwards the call to the normal GetMultiItemDefs or CItemMultiCustom::GetMultiItemDefs as appropriate, replacing the need to remember to implement the same checks everywhere that needs to get the multi items.

+++ 0.56.b/common/CGrayData.cpp
@@ -321,7 +321,7 @@ CGrayTerrainInfo::CGrayTerrainInfo( TERRAIN_TYPE id )
 //*********************************************8
 // -CGrayMulti
 
-int CGrayMulti::Load( MULTI_TYPE id )
+size_t CGrayMulti::Load( MULTI_TYPE id )
 {
 	ADDTOCALLSTACK("CGrayMulti::Load");
 	// Just load the whole thing.
@@ -364,7 +364,7 @@ int CGrayMulti::Load( MULTI_TYPE id )
 			}
 
 			// copy to new format
-			for (int i = 0; i < m_iItemQty; i++)
+			for (size_t i = 0; i < m_iItemQty; i++)
 			{
 				m_pItems[i].m_wTileID = pItems[i].m_wTileID;
 				m_pItems[i].m_dx = pItems[i].m_dx;
