+++ 0.56.b/graysvr/graysvr.cpp
@@ -284,7 +284,7 @@ LPCTSTR g_szServerDescription =
 #endif
 	" by www.sphereserver.com";
 
-int CObjBase::sm_iCount = 0;	// UID table.
+size_t CObjBase::sm_iCount = 0;	// UID table.
 LONGLONG llTimeProfileFrequency = 1000;	// time profiler
 
 // game servers stuff.
